+++ Pacman Remade/Pacman.h
@@ -21,23 +21,23 @@ public:
     void update(double deltaTime);
     void render(SDL_Renderer* renderer);
 
-    int getX() const { return m_x; }
-    int getY() const { return m_y; }
+    double getX() const { return m_x; }
+    double getY() const { return m_y; }
     int getSize() const { return m_size; }
     double getSpeed() const { return m_speed; }
     Direction getDirection() const { return m_direction; }
     SpriteSheet* getSpriteSheet() const { return m_spriteSheet; }
     Maze* getMaze() const { return m_maze; }
-    int getTileX() const { return m_x / m_maze->getTileSize(); }
-    int getTileY() const { return m_y / m_maze->getTileSize(); }
+    int getTileX() const { return static_cast<int>(m_x / m_maze->getTileSize()); }
+    int getTileY() const { return static_cast<int>(m_y / m_maze->getTileSize()); }
 
-    void setX(int x) { m_x = x; }
-    void setY(int y) { m_y = y; }
+    void setX(double x) { m_x = x; }
+    void setY(double y) { m_y = y; }
     void setDirection(Direction direction) { m_direction = direction; }
     void changeState(std::unique_ptr<State> newState);
 
 private:
-    int m_x, m_y;
+    double m_x, m_y;
     int m_size;
     Direction m_direction;
     double m_speed;
